app/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenTelemetry - Core
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# OpenTelemetry - Instrumentación
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-logging==0.43b0

# OpenTelemetry - Exporters
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# AI/ML
openai==1.10.0

# Utils
python-dotenv==1.0.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

